requirements.txt
numpy==1.26.4
pandas==2.2.2
scipy==1.14.0
scikit-learn==1.5.1
matplotlib==3.9.1
seaborn==0.13.2
tqdm==4.66.4
umap-learn==0.5.6
scanpy==1.10.2
torch==2.4.0
